Checking

+++ src/main/java/za/ac/cput/hospitalmanagementsystem/repository/AppointmentRepository.java
@@ -6,7 +6,7 @@ import za.ac.cput.hospitalmanagementsystem.domain.Appointment;
 /**
  * Created by student on 2015/05/01.
  */
-public interface AppointmentRepository extends CrudRepository<Appointment, Integer> {
+public interface AppointmentRepository extends CrudRepository<Appointment, Long> {
 
     public Appointment findByCode(String code);
