+++ src/main/java/za/ac/cput/hospitalmanagementsystem/repository/PatientRepository.java
@@ -6,7 +6,7 @@ import za.ac.cput.hospitalmanagementsystem.domain.Patient;
 /**
  * Created by student on 2015/05/01.
  */
-public interface PatientRepository extends CrudRepository<Patient, Integer> {
+public interface PatientRepository extends CrudRepository<Patient, Long> {
 
     public Patient findByCode(String code);
