+++ src/main/java/za/ac/cput/hospitalmanagementsystem/repository/WardRepository.java
@@ -6,7 +6,7 @@ import za.ac.cput.hospitalmanagementsystem.domain.Ward;
 /**
  * Created by student on 2015/05/01.
  */
-public interface WardRepository extends CrudRepository<Ward, Integer> {
+public interface WardRepository extends CrudRepository<Ward, Long> {
 
     public Ward findByCode(String code);
